+++ examples/data-sources/graylog_event_entity_types/data-source.tf
@@ -1,0 +1,5 @@
+data "graylog_event_entity_types" "available" {}
+
+output "event_processors" {
+  value = data.graylog_event_entity_types.available.processor_types
+}