+++ examples/resources/graylog_lookup_cache/resource.tf
@@ -1,9 +1,24 @@
-resource "graylog_lookup_cache" "example" {
-  title       = "Terraform Lookup Cache"
-  name        = "terraform-lookup-cache"
+resource "graylog_lookup_cache" "none" {
+  title       = "Terraform Null Cache"
+  name        = "terraform-lookup-cache-none"
   description = "Managed by Terraform"
 
   config = {
     type = "none"
+  }
+}
+
+resource "graylog_lookup_cache" "guava" {
+  title       = "Terraform Guava Cache"
+  name        = "terraform-lookup-cache-guava"
+  description = "Managed by Terraform"
+
+  config = {
+    type                     = "guava_cache"
+    max_size                 = 1000
+    expire_after_access      = 60
+    expire_after_access_unit = "SECONDS"
+    expire_after_write       = 0
+    ignore_null              = false
   }
 }